refactored default rules to already be added in the rules service
