+++ src/permissions/rules/rules.service.spec.ts
@@ -1,11 +1,16 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { DocumentRule, RulesService } from './rules.service';
-import { User } from '../../session/session/user-auth.dto';
+import { COUCHDB_USER_DOC, User } from '../../session/session/user-auth.dto';
 import { HttpService } from '@nestjs/axios';
 import { firstValueFrom, of, throwError } from 'rxjs';
 import { Permission } from './permission';
 import { CouchProxyController } from '../../replication/couch-proxy/couch-proxy.controller';
 import { ConfigService } from '@nestjs/config';
+import { DatabaseDocument } from '../../replication/couch-proxy/couchdb-dtos/bulk-docs.dto';
+import {
+  detectDocumentType,
+  DocumentAbility,
+} from '../permission/permission.service';
 
 describe('RulesService', () => {
   let service: RulesService;
@@ -22,7 +27,7 @@ describe('RulesService', () => {
         { action: 'read', subject: 'Aser' },
         { action: 'read', subject: 'Child', inverted: true },
       ],
-      admin_app: [{ action: 'manage', subject: 'Child' }],
+      admin_app: [{ action: 'manage', subject: 'all' }],
     });
     mockHttpService = {
       get: () => of({ data: testPermission }),
@@ -66,29 +71,78 @@ describe('RulesService', () => {
 
     await firstValueFrom(service.loadRules());
 
-    expect(service.getRulesForUser(new User('some-user', []))).toEqual([
-      {
-        subject: 'all',
-        action: 'manage',
-      },
-    ]);
+    expect(service.getRulesForUser(new User('some-user', []))).toContainEqual({
+      subject: 'all',
+      action: 'manage',
+    });
   });
 
-  it('should only return the rules for the passed user roles', () => {
+  it('should return the rules for every passed user role', () => {
     let result = service.getRulesForUser(new User('normalUser', ['user_app']));
 
-    expect(result).toEqual(userRules);
+    userRules.forEach((rule) => expect(result).toContain(rule));
+    adminRules.forEach((rule) => expect(result).not.toContain(rule));
 
     result = service.getRulesForUser(
       new User('superUser', ['user_app', 'admin_app']),
     );
-    expect(result).toEqual(userRules.concat(adminRules));
+    userRules
+      .concat(adminRules)
+      .forEach((rule) => expect(result).toContain(rule));
   });
 
   it('should not fail if no rules exist for a given role', () => {
     const result = service.getRulesForUser(
       new User('specialUser', ['user_app', 'manager_app']),
     );
-    expect(result).toEqual(userRules);
+    userRules.forEach((rule) => expect(result).toContain(rule));
+  });
+
+  it('should return a ability that does not allow to modify the permission document', () => {
+    const rules = service.getRulesForUser(new User('someUser', ['admin_app']));
+    const ability = new DocumentAbility(rules, {
+      detectSubjectType: detectDocumentType,
+    });
+
+    const permissionDoc: Permission = {
+      _id: `Permission:${Permission.DOC_ID}`,
+      _rev: 'someRev',
+      rulesConfig: {},
+    };
+    expect(ability.cannot('create', permissionDoc)).toBe(true);
+    expect(ability.cannot('update', permissionDoc)).toBe(true);
+    expect(ability.cannot('delete', permissionDoc)).toBe(true);
+    expect(ability.can('read', permissionDoc)).toBe(true);
+  });
+
+  it('should return a ability where normal users can only read their own user doc and update their password', () => {
+    const testUser = new User('someUser', []);
+    const rules = service.getRulesForUser(testUser);
+    const ability = new DocumentAbility(rules, {
+      detectSubjectType: detectDocumentType,
+    });
+
+    const userDoc: DatabaseDocument = {
+      _id: `${COUCHDB_USER_DOC}:${testUser.name}`,
+      _rev: 'someRev',
+      name: testUser.name,
+    };
+
+    const otherUserDoc: DatabaseDocument = {
+      _id: `${COUCHDB_USER_DOC}:otherUser`,
+      _rev: 'otherRev',
+      name: 'otherUser',
+    };
+
+    expect(ability.can('read', userDoc)).toBe(true);
+    expect(ability.can('update', userDoc, 'password')).toBe(true);
+    expect(ability.cannot('update', userDoc, 'roles')).toBe(true);
+    expect(ability.cannot('create', userDoc)).toBe(true);
+    expect(ability.cannot('delete', userDoc)).toBe(true);
+    expect(ability.cannot('read', otherUserDoc)).toBe(true);
+    expect(ability.cannot('update', otherUserDoc)).toBe(true);
+    expect(ability.cannot('update', otherUserDoc, 'password')).toBe(true);
+    expect(ability.cannot('create', otherUserDoc)).toBe(true);
+    expect(ability.cannot('delete', otherUserDoc)).toBe(true);
   });
 });
