+++ src/permissions/rules/rules.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { RawRuleOf } from '@casl/ability';
 import { DocumentAbility } from '../permission/permission.service';
-import { User } from '../../session/session/user-auth.dto';
+import { COUCHDB_USER_DOC, User } from '../../session/session/user-auth.dto';
 import { HttpService } from '@nestjs/axios';
 import { CouchDBInteracter } from '../../utils/couchdb-interacter';
 import { ConfigService } from '@nestjs/config';
@@ -17,6 +17,11 @@ export type DocumentRule = RawRuleOf<DocumentAbility>;
  */
 @Injectable()
 export class RulesService extends CouchDBInteracter {
+  private readonly permissionWriteRestriction: DocumentRule = {
+    subject: 'Permission',
+    action: ['create', 'update', 'delete'],
+    inverted: true,
+  };
   private permission: Permission;
 
   constructor(httpService: HttpService, configService: ConfigService) {
@@ -41,13 +46,42 @@ export class RulesService extends CouchDBInteracter {
    * @returns DocumentRule[] rules that are related to the user
    */
   getRulesForUser(user: User): DocumentRule[] {
+    const userRules = this.getRulesForRoles(user.roles);
+    return userRules.concat(...this.getDefaultRules(user));
+  }
+
+  private getRulesForRoles(roles: string[]): DocumentRule[] {
     if (this.permission && this.permission.rulesConfig) {
-      return user.roles
+      return roles
         .filter((role) => this.permission.rulesConfig.hasOwnProperty(role))
         .map((role) => this.permission.rulesConfig[role])
         .flat();
     } else {
       return [{ subject: 'all', action: 'manage' }];
     }
+  }
+
+  private getDefaultRules(user: User): DocumentRule[] {
+    const presetRules: DocumentRule[] = [this.permissionWriteRestriction];
+    if (!user.roles.includes('_admin')) {
+      // normal users can only read their own user object and update their password
+      presetRules.push({
+        subject: COUCHDB_USER_DOC,
+        action: ['manage'],
+        inverted: true,
+      });
+      presetRules.push({
+        subject: COUCHDB_USER_DOC,
+        action: 'read',
+        conditions: { name: user.name },
+      });
+      presetRules.push({
+        subject: COUCHDB_USER_DOC,
+        action: 'update',
+        fields: 'password',
+        conditions: { name: user.name },
+      });
+    }
+    return presetRules;
   }
 }
